requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
h11==0.13.0
idna==3.3
outcome==1.2.0
pycparser==2.21
pymongo==4.1.1
pyOpenSSL==22.0.0
PySocks==1.7.1
pyTelegramBotAPI==4.6.0
python-dotenv==0.20.0
requests==2.28.0
selenium==4.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0

asyncssh~=2.13.0